requirements.txt
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
pandas==2.1.3
pgmpy==0.1.24
Pillow==10.0.1
Pillow==10.1.0
Requests==2.31.0
scikit_learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
streamlit==1.27.2
streamlit_lottie==0.0.5
